requirements.txt
fastapi
uvicorn
google-genai
kerykeion
geopy
timezonefinder
pytz
pydantic
python-multipart
python-dotenv